Attempt to merge some codeFrom originalAndCondense
