+++ example.js
@@ -1,7 +1,7 @@
 //Used for testing various functions; used only in development, plz ignore
 
 const pl = require("./protocollib.js");
-pl.debug = true;
+pl.setDebug(true);
 const http = require("http");
 
 var server = http.createServer(()=>{});
